# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: string;
  name: string;
  age: number;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  photo_url?: string | null;
  sparks_balance: number;
  integrity_score: number;
  personality_tags?: string[];
  is_verified?: boolean;
  created_at?: string;
}

export interface ProfileUpdate {
  name?: string;
  age?: number;
  bio?: string;
  city?: string;
  photo_url?: string;
  personality_tags?: string[];
}

export interface AudioPrompt {
  id: string;
  user_id: string;
  prompt_question: string;
  audio_url: string;
  /** @nullable */
  duration_seconds?: number | null;
  created_at?: string;
}

export interface AudioPromptInput {
  prompt_question: string;
  audio_url: string;
  duration_seconds?: number;
}

export interface DiscoverCard {
  match_id: string;
  user_id?: string;
  name?: string;
  age?: number;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  distance_km?: number | null;
  /** @nullable */
  photo_url?: string | null;
  photo_revealed?: boolean;
  audio_prompts: AudioPrompt[];
  personality_tags?: string[];
  expires_at: string;
  status?: string;
}

export interface Match {
  id: string;
  matched_user?: UserProfile;
  status: string;
  photo_revealed: boolean;
  chat_unlocked: boolean;
  message_count: number;
  message_limit: number;
  expires_at: string;
  /** @nullable */
  matched_at?: string | null;
  created_at?: string;
}

export interface Message {
  id: string;
  match_id: string;
  sender_id: string;
  content: string;
  is_read: boolean;
  is_unsent: boolean;
  sent_at: string;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export interface DailyEarnOption {
  type: string;
  label: string;
  amount: number;
  claimed: boolean;
}

export interface SparksBalance {
  balance: number;
  earned_this_week: number;
  spent_this_week: number;
  free_keys_remaining?: number;
  free_unsends_remaining?: number;
  daily_earn_available?: DailyEarnOption[];
}

export interface SparksTransaction {
  id: string;
  amount: number;
  reason: string;
  balance_after: number;
  created_at: string;
}

export interface SparksBundle {
  id: string;
  name: string;
  sparks: number;
  price_usd: number;
  /** @nullable */
  label: string | null;
  /** @nullable */
  description?: string | null;
}

export interface EarnClaimInput {
  claim_type: string;
}

export interface PurchaseInput {
  bundle_id: string;
}