# lib/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  title: Api
  version: "1.0.0"
servers:
  - url: /api
paths:
...
  /profile/me:
    get:
      operationId: getMyProfile
      summary: Get current user profile
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    put:
      operationId: updateMyProfile
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /profile/{userId}:
    get:
      operationId: getUserProfile
      summary: Get another user's profile (for match view)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found

  # ─── Audio Prompts ──────────────────────────────────────────────────────────
  /prompts:
    get:
      operationId: getMyPrompts
      summary: Get my audio prompt responses
      responses:
        "200":
          description: List of prompts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AudioPrompt"
    post:
      operationId: createPrompt
      summary: Add an audio prompt response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AudioPromptInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AudioPrompt"

  # ─── Discover ──────────────────────────────────────────────────────────────
...
  /matches/{matchId}:
    get:
      operationId: getMatch
      summary: Get a single match
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Match detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
        "404":
          description: Not found
  /matches/{matchId}/accept:
    post:
      operationId: acceptMatch
      summary: Accept a match blind (no photo reveal)
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Accepted match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /matches/{matchId}/reveal:
    post:
      operationId: revealMatchPhoto
      summary: Reveal match photo early (costs 2 Sparks)
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Match with photo revealed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
        "402":
          description: Insufficient Sparks
  /matches/{matchId}/extend:
    post:
      operationId: extendMatch
      summary: Extend match timer by 2 hours (costs 1 Spark)
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Match with extended timer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
        "402":
          description: Insufficient Sparks
  /matches/{matchId}/unlock-chat:
    post:
      operationId: unlockChat
      summary: Use a chat key to unlock messaging (1 free/day, else 5 Sparks)
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chat unlocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
        "402":
          description: Insufficient Sparks or keys
  /matches/{matchId}/stretch:
    post:
      operationId: stretchMessages
      summary: Add 10 more messages to a chat (costs 3 Sparks)
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Match with increased message limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
        "402":
          description: Insufficient Sparks

  # ─── Messages ──────────────────────────────────────────────────────────────
  /matches/{matchId}/messages:
    get:
      operationId: getMessages
      summary: Get messages for a match
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Message list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      summary: Send a message in a match
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "403":
          description: Chat locked or message limit reached
  /messages/{messageId}/unsend:
    post:
      operationId: unsendMessage
      summary: Unsend a message within 60s (costs 2 Sparks after 1 free/day)
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Message unsent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "402":
          description: Insufficient Sparks
        "410":
          description: Unsend window expired (>60s)

  # ─── Sparks ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, name, age, sparks_balance, integrity_score]
      properties:
        id:
          type: string
        name:
          type: string
        age:
          type: integer
        bio:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        photo_url:
          type: ["string", "null"]
        sparks_balance:
          type: integer
        integrity_score:
          type: integer
        personality_tags:
          type: array
          items:
            type: string
        is_verified:
          type: boolean
        created_at:
          type: string
    ProfileUpdate:
      type: object
      properties:
        name:
          type: string
        age:
          type: integer
        bio:
          type: string
        city:
          type: string
        photo_url:
          type: string
        personality_tags:
          type: array
          items:
            type: string
    AudioPrompt:
      type: object
      required: [id, user_id, prompt_question, audio_url]
      properties:
        id:
          type: string
        user_id:
          type: string
        prompt_question:
          type: string
        audio_url:
          type: string
        duration_seconds:
          type: ["integer", "null"]
        created_at:
          type: string
    AudioPromptInput:
      type: object
      required: [prompt_question, audio_url]
      properties:
        prompt_question:
          type: string
        audio_url:
          type: string
        duration_seconds:
          type: integer
...
    Match:
      type: object
      required: [id, status, expires_at, chat_unlocked, message_count, message_limit, photo_revealed]
      properties:
        id:
          type: string
        matched_user:
          $ref: "#/components/schemas/UserProfile"
        status:
          type: string
        photo_revealed:
          type: boolean
        chat_unlocked:
          type: boolean
        message_count:
          type: integer
        message_limit:
          type: integer
        expires_at:
          type: string
        matched_at:
          type: ["string", "null"]
        created_at:
          type: string
    Message:
      type: object
      required: [id, match_id, sender_id, content, sent_at, is_read, is_unsent]
      properties:
        id:
          type: string
        match_id:
          type: string
        sender_id:
          type: string
        content:
          type: string
        is_read:
          type: boolean
        is_unsent:
          type: boolean
        sent_at:
          type: string
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
...