# lib/api-zod/src/generated/types/audioPrompt.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface AudioPrompt {
  id: string;
  user_id: string;
  prompt_question: string;
  audio_url: string;
  /** @nullable */
  duration_seconds?: number | null;
  created_at?: string;
}