# lib/api-zod/src/generated/types/audioPromptInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface AudioPromptInput {
  prompt_question: string;
  audio_url: string;
  duration_seconds?: number;
}