# lib/api-zod/src/generated/types/match.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
import type { UserProfile } from './userProfile';

export interface Match {
  id: string;
  matched_user?: UserProfile;
  status: string;
  photo_revealed: boolean;
  chat_unlocked: boolean;
  message_count: number;
  message_limit: number;
  expires_at: string;
  /** @nullable */
  matched_at?: string | null;
  created_at?: string;
}