# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface Message {
  id: string;
  match_id: string;
  sender_id: string;
  content: string;
  is_read: boolean;
  is_unsent: boolean;
  sent_at: string;
}