# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface ProfileUpdate {
  name?: string;
  age?: number;
  bio?: string;
  city?: string;
  photo_url?: string;
  personality_tags?: string[];
}