# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface UserProfile {
  id: string;
  name: string;
  age: number;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  photo_url?: string | null;
  sparks_balance: number;
  integrity_score: number;
  personality_tags?: string[];
  is_verified?: boolean;
  created_at?: string;
}